02/04/2018 -- pcesCommon / pcesGerador -- [*] Correção do tipo do campo qtdHrsSem de Integer para Double.    Por: Renato E -- pcesGerador -- [*] https://www.projetoacbr.com.br/forum/topic/39746-erro-ao-assinar-xml-no-esocial/?do=findComment&comment=279903    Por: Leandro JLM -- pcesGerador / pcesS2306 -- [*] A geração do grupo Remuneração nos eventos 2200 e 2206 é obrigatória, já nos eventos 2300 e 2306     é opcional, na unit pcesGerador foi removido o IF e na unit 2306 foi criado a procedure para gerar o     grupo.    Por: Italo Jurisato Junior -- pcesS2306 / pcesS2399 / pcesS2400 / pcesS3000 -- [+] Implementação da leitura do arquivo INI referente aos eventos S-2306,     S-2399, S-2400 e S-3000.    Por: Italo Jurisato Junior
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@14939 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrDFe/ACBreSocial/PCNeSocial/pcesCommon.pas
@@ -560,7 +560,7 @@ type
 
   THorContratual = class(TPersistent)
   private
-    FQtdHrsSem: integer;
+    FQtdHrsSem: Double;
     FTpJornada: tpTpJornada;
     FDscTpJorn: string;
     FTMPParc: tpTmpParc;
@@ -569,7 +569,7 @@ type
     constructor Create;
     destructor Destroy; override;
 
-    property QtdHrsSem: integer read FQtdHrsSem write FQtdHrsSem;
+    property QtdHrsSem: Double read FQtdHrsSem write FQtdHrsSem;
     property TpJornada: tpTpJornada read FTpJornada write FTpJornada;
     property DscTpJorn: string read FDscTpJorn write FDscTpJorn;
     property tmpParc: tpTmpParc read FTMPParc write FTMPParc;
